Disallow multiple cities to be selected
This ends up making the filter state a bit confusing, so it's easiest just to disable it for now.

Really, cities should be selectable via a radio button or a select dropdown if we want to go this way.

I'll need to revisit the filter logic to figure out how to allow multiple cities to be selected at once.